+++ lib/cpp/src/thrift/concurrency/Thread.h
@@ -39,7 +39,7 @@ class Thread;
 class Runnable {
 
 public:
-  virtual ~Runnable()= default;;
+  virtual ~Runnable() = default;
   virtual void run() = 0;
 
   /**